+++ SurfaceTopography/Support/UnitConversion.py
@@ -31,6 +31,17 @@ voltage_units = {'GV': 1e9, 'MV': 1e6, 'kV': 1000.0, 'V': 1.0, 'mV': 1e-3, 'µV'
 
 units = dict(length=length_units, voltage=voltage_units)
 
+mangle_units = {
+    'A': 'Å',
+    'NanoMeters': 'nm',  # Zygo DATX
+    'μm': 'µm',
+    'um': 'µm',
+    '~m': 'µm',
+    'MicroMeters': 'µm',  # Zygo DATX
+    'MilliMeters': 'mm',  # Zygo DATX
+    'Meters': 'm',  # Zygo DATX
+}
+
 
 def is_length_unit(s):
     return s in length_units.keys()
@@ -101,10 +112,8 @@ def mangle_length_unit_utf8(unit):
         unit = unit.decode('utf-8').strip()
     if unit == '':
         return None
-    elif unit == 'A':
-        return 'Å'
-    elif unit == 'μm' or unit == 'um' or unit == '~m':
-        return 'µm'
+    elif unit in mangle_units:
+        return mangle_units[unit]
     else:
         return unit
 
@@ -127,10 +136,8 @@ def mangle_length_unit_ascii(unit):
     unit = unit.strip()
     if unit == '':
         return None
-    elif unit == 'Å':
-        return 'A'
-    elif unit == 'μm' or unit == 'µm' or unit == '~m':
-        return 'um'
+    elif unit in mangle_units:
+        return mangle_units[unit]
     else:
         return unit